+++ clash-prelude/src/Clash/Explicit/Testbench.hs
@@ -52,9 +52,9 @@ import Clash.Promoted.Symbol (SSymbol (..))
 import Clash.Explicit.Signal
   (Clock, Reset, System, Signal, clockPeriod, toEnable, fromList, register,
   unbundle, unsafeSynchronizer, veryUnsafeSynchronizer)
-import Clash.Signal.Internal (Clock (..))
+import Clash.Signal.Internal (Clock (..), Reset (..))
 import Clash.Signal
-  (mux, DomainResetKind, ResetKind(Asynchronous), KnownDomain, Reset(..),
+  (mux, DomainResetKind, ResetKind(Asynchronous), KnownDomain,
   Enable)
 import Clash.Sized.Index     (Index)
 import Clash.Sized.Internal.BitVector